Split keys_spec.rb into user_, client_keys_spec.rb
